Changed some ints to shorts.

+++ basketballGAME/basketballAPI/models/StatsType.cs
@@ -5,7 +5,7 @@ namespace basketballAPI.models;
 
 public partial class StatsType
 {
-    public int StatId { get; set; }
+    public short StatId { get; set; }
 
     public string? StatName { get; set; }
